refactor(agent-output): extract MarkdownTheme from the renderer
Fence-body handlers land in sibling files and need the theme type, which
was a private nested class. Pure move: same Resolve, same Find, same
fallbacks, no behavior change.
